requirements.txt
altgraph==0.17.4
asgiref==3.8.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.4
distlib==0.3.9
Django==5.1.5
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
filelock==3.16.1
idna==3.10
ms-identity-python @ https://github.com/azure-samples/ms-identity-python/archive/refs/heads/0.9.zip#sha256=4322333417c2cb5c89056e5a8feca3c750a2753b67c558e85cb97ac4ff8b0fda
msal==1.32.3
mysqlclient==2.2.7
packaging==24.2
pefile==2023.2.7
platformdirs==4.3.6
pycparser==2.22
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
PyJWT==2.9.0
python-dotenv==0.21.1
pywin32-ctypes==0.2.3
requests==2.32.4
setuptools==75.8.1
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.4.0
virtualenv==20.28.1
